decompressor: use a size_t for currentThread
This is the correct type for what it is doing (array access).
Switching it also avoids a signed/unsigned comparison warning.

+++ c-ext/decompressor.c
@@ -1099,7 +1099,7 @@ ZstdBufferWithSegmentsCollection* decompress_from_framesources(ZstdDecompressor*
 	ZstdBufferWithSegmentsCollection* result = NULL;
 	FramePointer* framePointers = frames->frames;
 	unsigned long long workerBytes = 0;
-	int currentThread = 0;
+	size_t currentThread = 0;
 	Py_ssize_t workerStartOffset = 0;
 	POOL_ctx* pool = NULL;
 	WorkerState* workerStates = NULL;
